+++ src/main/java/com/proyecto/MarketDelivery/model/Agenda.java
@@ -1,6 +1,6 @@
 package com.proyecto.MarketDelivery.model;
 
-import com.proyecto.MarketDelivery.model.Servicio;
+
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
@@ -10,7 +10,7 @@ public class Agenda {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "Id_Agenda")
-    private Integer id;
+    private Long id;
 
     @Column(name = "Disponibilidad", nullable = false)
     private LocalDateTime disponibilidad;
@@ -20,8 +20,8 @@ public class Agenda {
     private Servicio servicio;
 
     // Getters y Setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public LocalDateTime getDisponibilidad() { return disponibilidad; }
     public void setDisponibilidad(LocalDateTime disponibilidad) { this.disponibilidad = disponibilidad; }
