+++ src/main/java/com/proyecto/MarketDelivery/repository/DetalleFacturaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DetalleFacturaRepository extends JpaRepository<DetalleFactura, Integer> {
+public interface DetalleFacturaRepository extends JpaRepository<DetalleFactura, Long> {
 }
