+++ src/main/java/com/proyecto/MarketDelivery/repository/PqrRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PqrRepository extends JpaRepository<Pqr, Integer> {
+public interface PqrRepository extends JpaRepository<Pqr, Long> {
 }
